rename component and test files
capitalize filenames, use .jsx extension
