+++ test/frontend/components/site/SitePublishedFilesTable.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 import LoadingIndicator from '../../../../frontend/components/LoadingIndicator';
-import SitePublishedFilesTable from '../../../../frontend/components/site/sitePublishedFilesTable';
+import SitePublishedFilesTable from '../../../../frontend/components/site/SitePublishedFilesTable';
 
 
 describe('<SitePublishedFilesTable/>', () => {
